[AI Tutor] Corrected migration for ai tutor interaction feedback (#57946)

+++ dashboard/db/schema.rb
@@ -41,7 +41,7 @@ ActiveRecord::Schema.define(version: 2024_04_03_074309) do
 
   create_table "ai_tutor_interaction_feedbacks", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.bigint "ai_tutor_interaction_id", null: false
-    t.bigint "user_id", null: false
+    t.integer "user_id", null: false
     t.boolean "thumbs_up"
     t.boolean "thumbs_down"
     t.datetime "created_at", precision: 6, null: false
